Refactor: Removed region from location.
Find me button returned city and region which were saved together as location.
Region is moved now to avoid confusion in the database in case the user enters city alone.
For exmaple, if user in Seattle and searches Seattle, location is 'seattle'.
If using find me it will be 'seattle washington'. So depending on how search is done, info retreival from database may differ.