Bug and Refactor: Fixed bug which resulted in wring buttons in results page in some cases.
Before, if the user was logged in and performed a search, the code wouldn't check what they already decided to go to,
so the button 'I want to go' appeared everywhere.
This is because post request on /results didn't check for that.
get and post request on /results now call the function renderResults which is called when user is logged in.
The function now performs that check and sets the isGoing property correctly which in turn decides which buttons render.